+++ packages/kysely/src/themes/index.ts
@@ -1,0 +1,16 @@
+import { create } from "./create";
+import { Kysely } from "kysely";
+import { remove } from "./remove";
+import { get } from "./get";
+import { update } from "./update";
+import { ThemesAdapter } from "@authhero/adapter-interfaces";
+import { Database } from "../db";
+
+export function createConnectionsAdapter(db: Kysely<Database>): ThemesAdapter {
+  return {
+    create: create(db),
+    get: get(db),
+    remove: remove(db),
+    update: update(db),
+  };
+}