+++ packages/kysely/src/themes/remove.ts
@@ -1,0 +1,14 @@
+import { Kysely } from "kysely";
+import { Database } from "../db";
+
+export function remove(db: Kysely<Database>) {
+  return async (tenant_id: string, themeId: string): Promise<boolean> => {
+    const result = await db
+      .deleteFrom("themes")
+      .where("themes.tenant_id", "=", tenant_id)
+      .where("themes.theme_id", "=", themeId)
+      .executeTakeFirst();
+
+    return result.numDeletedRows > 0;
+  };
+}